Extracted presentation logic out into MostViewedGrid
