+++ packages/frontend/web/components/MostViewed/MostViewed.tsx
@@ -1,25 +1,19 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { css, cx } from 'emotion';
-import { textSans, headline, palette } from '@guardian/src-foundations';
+import { headline, palette } from '@guardian/src-foundations';
 import {
     desktop,
-    tablet,
     leftCol,
     wide,
-    phablet,
     mobileLandscape,
 } from '@guardian/pasteup/breakpoints';
-import { screenReaderOnly } from '@guardian/pasteup/mixins';
-import { Container, BigNumber } from '@guardian/guui';
+import { Container } from '@guardian/guui';
 import { namedAdSlotParameters } from '@frontend/model/advertisement';
 import { AdSlot, labelStyles } from '@frontend/web/components/AdSlot';
-import ClockIcon from '@guardian/pasteup/icons/clock.svg';
-import { PulsingDot } from '@frontend/web/components/PulsingDot';
-import { QuoteIcon } from '@frontend/web/components/QuoteIcon';
-import { pillarPalette } from '@frontend/lib/pillars';
 import { OutbrainContainer } from '@frontend/web/components/Outbrain';
-
 import { useApi } from '@frontend/web/components/lib/api';
+
+import { MostViewedGrid } from './MostViewedGrid';
 
 const container = css`
     padding-top: 3px;
@@ -74,160 +68,6 @@ const heading = css`
     }
 `;
 
-const listContainer = css`
-    max-width: 460px;
-
-    ${leftCol} {
-        margin-left: 160px;
-    }
-
-    ${wide} {
-        margin-left: 230px;
-    }
-`;
-
-const list = css`
-    margin-top: 12px;
-
-    ${tablet} {
-        border-top: 1px solid ${palette.neutral[86]};
-        width: 620px;
-        min-height: 300px;
-        column-width: 300px;
-        column-gap: 20px;
-        column-fill: balance;
-        column-count: 2;
-    }
-`;
-
-const hideList = css`
-    display: none;
-`;
-
-const listItem = css`
-    position: relative;
-    box-sizing: border-box;
-    padding-top: 4px;
-    padding-bottom: 24px;
-
-    &:before {
-        position: absolute;
-        top: 0;
-        right: 10px;
-        left: 0;
-        content: '';
-        display: block;
-        width: 100%;
-        height: 1px;
-        background-color: ${palette.neutral[86]};
-    }
-
-    :first-of-type {
-        &:before {
-            display: none;
-        }
-    }
-
-    &:after {
-        content: '';
-        display: block;
-        clear: both;
-    }
-
-    ${tablet} {
-        padding-top: 3px;
-        padding-bottom: 0;
-        min-height: 72px;
-    }
-
-    ${desktop} {
-        height: 100%;
-        display: inline-block;
-        width: 100%;
-
-        :nth-of-type(6) {
-            &:before {
-                display: none;
-            }
-        }
-    }
-`;
-
-const bigNumber = css`
-    float: left;
-    margin-top: 3px;
-    fill: ${palette.neutral[7]};
-`;
-
-const headlineHeader = css`
-    margin-top: -4px;
-    margin-left: 70px;
-    padding-top: 2px;
-    padding-bottom: 2px;
-    word-wrap: break-word;
-    overflow: hidden;
-`;
-
-const headlineLink = css`
-    text-decoration: none;
-    color: ${palette.neutral[7]};
-    font-weight: 500;
-    ${headline({ level: 1 })};
-`;
-
-const tabsContainer = css`
-    border-bottom: 1px solid ${palette.neutral[86]};
-
-    &::after {
-        content: '';
-        display: block;
-        clear: left;
-
-        ${tablet} {
-            display: none;
-        }
-    }
-
-    ${tablet} {
-        border-bottom: 0;
-    }
-`;
-
-const listTab = css`
-    width: 50%;
-    float: left;
-    border-top: 3px solid ${palette.neutral[93]};
-    background-color: ${palette.neutral[93]};
-
-    ${phablet} {
-        width: 230px;
-    }
-`;
-
-const selectedListTab = css`
-    background-color: ${palette.neutral[100]};
-`;
-
-const tabButton = css`
-    ${headline({ level: 1 })};
-    margin: 0;
-    border: 0;
-    background: transparent;
-    padding-left: 10px;
-    padding-right: 6px;
-    padding-top: 4px;
-    text-align: left;
-    text-decoration: none;
-    font-weight: 600;
-    min-height: 36px;
-    display: block;
-    width: 100%;
-
-    &:hover {
-        cursor: pointer;
-    }
-`;
-
 const adSlotUnspecifiedWidth = css`
     .ad-slot {
         margin: 12px auto;
@@ -252,63 +92,7 @@ const mostPopularAdStyle = css`
     ${labelStyles};
 `;
 
-const liveKicker = (colour: string) => css`
-    color: ${colour};
-    font-weight: 700;
-
-    &::after {
-        content: '/';
-        display: inline-block;
-        font-weight: 900;
-        margin: 0 4px;
-    }
-`;
-
-function getColour(pillar: Pillar) {
-    return pillarPalette[pillar].main;
-}
-
-const oldArticleMessage = css`
-    ${textSans({ level: 1 })}
-    background: ${palette.yellow.main};
-    display: inline-block;
-    color: ${palette.neutral[7]};
-    margin: 4px 0 6px;
-    padding: 3px 5px;
-
-    svg {
-        fill: currentColor;
-    }
-
-    .embolden {
-        font-weight: bold;
-    }
-`;
-
-const oldClockWrapper = css`
-    margin-right: 3px;
-`;
-
-const AgeWarning: React.FC<{
-    ageWarning?: string;
-}> = ({ ageWarning }) => {
-    if (!ageWarning) {
-        return <></>;
-    }
-    return (
-        <div>
-            <div className={oldArticleMessage}>
-                <span className={oldClockWrapper}>
-                    <ClockIcon />
-                </span>
-                This article is more than{' '}
-                <span className="embolden">{ageWarning} old</span>
-            </div>
-        </div>
-    );
-};
-
-interface TrailType {
+export interface TrailType {
     url: string;
     linkText: string;
     isLiveBlog: boolean;
@@ -316,13 +100,13 @@ interface TrailType {
     pillar: Pillar;
 }
 
-interface TabType {
+export interface TabType {
     heading: string;
     trails: TrailType[];
 }
 
 interface Props {
-    sectionName?: string;
+    sectionName: string;
     config: ConfigType;
 }
 
@@ -336,10 +120,8 @@ function buildSectionUrl(sectionName?: string) {
 }
 
 export const MostViewed = ({ config, sectionName }: Props) => {
-    const [selectedTabIndex, setSelectedTabIndex] = useState<number>(0);
-
     const url = buildSectionUrl(sectionName);
-    const { data, error } = useApi(url);
+    const { data, error } = useApi<TabType[]>(url);
 
     if (error) {
         window.guardian.modules.raven.reportError(
@@ -351,138 +133,36 @@ export const MostViewed = ({ config, sectionName }: Props) => {
         );
     }
 
-    return (
-        <div className={`content-footer ${cx(adSlotUnspecifiedWidth)}`}>
-            <OutbrainContainer config={config} />
-            <Container borders={true} className={cx(articleContainerStyles)}>
-                <div
-                    className={cx(container, mostPopularAdStyle)}
-                    data-link-name={'most-viewed'}
-                    data-component={'most-viewed'}
+    if (data) {
+        return (
+            <div className={`content-footer ${cx(adSlotUnspecifiedWidth)}`}>
+                <OutbrainContainer config={config} />
+                <Container
+                    borders={true}
+                    className={cx(articleContainerStyles)}
                 >
-                    <h2 className={heading}>Most popular</h2>
-                    <div className={mostPopularBody}>
-                        <div className={listContainer}>
-                            {Array.isArray(data) && data.length > 1 && (
-                                <ul className={tabsContainer} role="tablist">
-                                    {(data || []).map(
-                                        (tab: TabType, i: number) => (
-                                            <li
-                                                className={cx(listTab, {
-                                                    [selectedListTab]:
-                                                        i === selectedTabIndex,
-                                                })}
-                                                role="tab"
-                                                aria-selected={
-                                                    i === selectedTabIndex
-                                                }
-                                                aria-controls={`tabs-popular-${i}`}
-                                                id={`tabs-popular-${i}-tab`}
-                                                key={`tabs-popular-${i}-tab`}
-                                            >
-                                                <button
-                                                    className={tabButton}
-                                                    onClick={() =>
-                                                        setSelectedTabIndex(i)
-                                                    }
-                                                >
-                                                    <span
-                                                        className={css`
-                                                            ${screenReaderOnly};
-                                                        `}
-                                                    >
-                                                        Most viewed{' '}
-                                                    </span>
-                                                    <span // tslint:disable-line:react-no-dangerous-html
-                                                        // "Across The Guardian" has a non-breaking space entity between "The" and "Guardian"
-                                                        dangerouslySetInnerHTML={{
-                                                            __html: tab.heading,
-                                                        }}
-                                                    />
-                                                </button>
-                                            </li>
-                                        ),
-                                    )}
-                                </ul>
-                            )}
-                            {(data || []).map((tab: TabType, i: number) => (
-                                <ol
-                                    className={cx(list, {
-                                        [hideList]: i !== selectedTabIndex,
-                                    })}
-                                    id={`tabs-popular-${i}`}
-                                    key={`tabs-popular-${i}`}
-                                    role="tabpanel"
-                                    aria-labelledby={`tabs-popular-${i}-tab`}
-                                    data-link-name={tab.heading}
-                                    data-testid={tab.heading}
-                                    data-link-context={`most-read/${sectionName}`}
-                                >
-                                    {(tab.trails || []).map(
-                                        (trail: TrailType, ii: number) => (
-                                            <li
-                                                className={listItem}
-                                                key={trail.url}
-                                                data-link-name={`${ii +
-                                                    1} | text`}
-                                            >
-                                                <span className={bigNumber}>
-                                                    <BigNumber index={ii + 1} />
-                                                </span>
-                                                <h2 className={headlineHeader}>
-                                                    <a
-                                                        className={headlineLink}
-                                                        href={trail.url}
-                                                        data-link-name={
-                                                            'article'
-                                                        }
-                                                    >
-                                                        {trail.isLiveBlog && (
-                                                            <span
-                                                                className={liveKicker(
-                                                                    getColour(
-                                                                        trail.pillar,
-                                                                    ),
-                                                                )}
-                                                            >
-                                                                <PulsingDot
-                                                                    colour={getColour(
-                                                                        trail.pillar,
-                                                                    )}
-                                                                />
-                                                                Live
-                                                            </span>
-                                                        )}
-                                                        {trail.pillar ===
-                                                            'opinion' && (
-                                                            <QuoteIcon
-                                                                colour={getColour(
-                                                                    trail.pillar,
-                                                                )}
-                                                            />
-                                                        )}
-                                                        {trail.linkText}
-                                                        <AgeWarning
-                                                            ageWarning={
-                                                                trail.ageWarning
-                                                            }
-                                                        />
-                                                    </a>
-                                                </h2>
-                                            </li>
-                                        ),
-                                    )}
-                                </ol>
-                            ))}
+                    <div
+                        className={cx(container, mostPopularAdStyle)}
+                        data-link-name={'most-viewed'}
+                        data-component={'most-viewed'}
+                    >
+                        <h2 className={heading}>Most popular</h2>
+                        <div className={mostPopularBody}>
+                            <MostViewedGrid
+                                data={data}
+                                sectionName={sectionName}
+                            />
+                            <AdSlot
+                                asps={namedAdSlotParameters('most-popular')}
+                                config={config}
+                                className={''}
+                            />
                         </div>
-                        <AdSlot
-                            asps={namedAdSlotParameters('most-popular')}
-                            config={config}
-                            className={''}
-                        />
                     </div>
-                </div>
-            </Container>
-        </div>
-    );
+                </Container>
+            </div>
+        );
+    }
+
+    return null;
 };
